Refactor again so classes can be used independent of web app.
