+++ snapassist/webapp/static/fonts/glyphicons-halflings-regular.woff2
