+++ snapassist/webapp/static/js/bootstrap.js
