+++ snapassist/webapp/static/js/npm.js
